refactor: move hello::Hello from lib to wire
