+++ lib/src/hello.rs
@@ -1,22 +1,9 @@
 use crate::common::{request, SerializableNone};
 use crate::error::ApiError;
+use lrzcc_wire::hello::Hello;
 use reqwest::blocking::Client;
 use reqwest::{Method, StatusCode};
-use serde::{Deserialize, Serialize};
-use std::fmt::Display;
 use std::rc::Rc;
-use tabled::Tabled;
-
-#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
-pub struct Hello {
-    pub message: String,
-}
-
-impl Display for Hello {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.message.as_str())
-    }
-}
 
 pub struct HelloApi {
     pub url: String,
