+++ wire/src/hello.rs
@@ -1,1 +1,14 @@
+use serde::{Deserialize, Serialize};
+use std::fmt::Display;
+use tabled::Tabled;
 
+#[derive(Clone, Debug, Deserialize, Serialize, Tabled)]
+pub struct Hello {
+    pub message: String,
+}
+
+impl Display for Hello {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.message.as_str())
+    }
+}
